+++ ntex-bytes/src/string.rs
@@ -1,7 +1,7 @@
 //! A UTF-8 encoded read-only string using Bytes as storage.
 use std::{borrow, fmt, hash, ops, slice, str};
 
-use crate::{Bytes, BytesMut, BytesVec};
+use crate::{Bytes, BytesVec, bytesmut::BytesMut};
 
 /// An immutable UTF-8 encoded string with [`Bytes`] as a storage.
 #[derive(Clone, Default, Eq, PartialOrd, Ord)]