+++ ntex-bytes/tests/test_buf.rs
@@ -41,7 +41,8 @@ fn test_bytes() {
 
 #[test]
 fn test_bytes_mut() {
-    let mut buf = BytesMut::from(b"hello");
+    let mut buf = BytesMut::with_capacity(30);
+    buf.extend_from_slice(b"hello");
     assert_eq!(Buf::remaining(&buf), 5);
     assert_eq!(Buf::chunk(&buf), b"hello");
 
@@ -59,7 +60,8 @@ fn test_bytes_mut() {
 
 #[test]
 fn test_bytes_mut_buf() {
-    let mut buf = BytesMut::from(b"hello");
+    let mut buf = BytesMut::with_capacity(30);
+    buf.extend_from_slice(b"hello");
 
     assert_eq!(bytes::buf::Buf::remaining(&buf), 5);
     assert_eq!(bytes::buf::Buf::chunk(&buf), b"hello");